+++ CHANGELOG.md
@@ -80,6 +80,7 @@ with the `YY.MINOR.MICRO` scheme.
 - Fixed that sentinel_loop was broken during upgrade from versions before 25.5.3.
 - Fixed ignoring `NULLS FIRST` and `NULLS LAST` in unnamed window queries with ordering.
 - Fixed metadata loss in queries with LIMIT clause in picodata admin.
+- Fixed invalid volatile flag for rust-implemented builtin functions.
 
 
 ## [25.5.5] - 2026-01-26